+++ Packages/net.fiakaiera.door/CHANGELOG.md
@@ -28,3 +28,8 @@
 - Merged duplicate code
 - Renamed assembly definition to match namespace
 - Added documentation links in package
+
+[v1.1.1](https://github.com/fiaKaiera/vpm-listing/releases/tag/door-1.1.1)
+- Opened and closed Points have been replaced with handles, no longer requiring two extra objects
+- Door object is now anchored at `0`, instead of `-1` for consistency for future `HingeDoor`
+- Handle object's layer is set to `Pickup`